backend/requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6
openapi-spec-validator>=0.7.1
PyYAML>=6.0.1
requests>=2.31.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.1
